requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
dataclasses-json==0.6.7
decorator==4.4.2
distro==1.9.0
dj-database-url==2.1.0
Django==5.1.4
django-anymail==10.3
django-browser-reload==1.12.0
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-payments==2.0.0
django-phonenumber-field==7.3.0
django-ratelimit==4.1.0
django-storages==1.14.2
django-tailwind==3.8.0
django-unfold==0.34.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.126.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.12.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.23.2
moviepy==1.0.3
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.58.1
openpyxl==3.1.2
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
orjson==3.10.12
packaging==24.2
pandas==2.2.0
phonenumberslite==8.13.52
pillow==10.3.0
proglog==0.1.10
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.10.1
pyparsing==3.2.0
pytz==2023.3.post1
PyYAML==6.0.2
regex==2024.11.6
replicate==0.33.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stripe==11.4.1
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.6.0
yarl==1.18.3
